Trait pallet_example::Trait

pub trait Trait: Trait {
    type Event: From<Event<Self>> + Into<Self::Event>;
}

Our pallet’s configuration trait. All our types and constants go in here. If the pallet is dependent on specific other pallets, then their configuration traits should be added to our implied traits list.

frame_system::Trait should always be included in our implied traits.

Associated Types

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

Loading content...